+++ CHANGELOG.md
@@ -7,21 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This changelog was started for release 0.0.3.
 
-## [0.0.3] - Unreleased
+## [0.0.3] - 21/11/2022
 
 ### Added
 
-- empty_ok_if key for validator
-- empty_ok_unless key for validator
+- empty_ok_if key for validator & templates
+- empty_ok_unless key for validator & templates
 - readme key for validator
 - unique key for validator
 - expected_rows key for templates
 - logs parameters for templates
+- na_ok key for validators & templates
+- skip_generation key for validators & templates
+- skip_validation key for validators & templates
 
 ### Fixed
 
 - Bug for setValidator when using number values
+- Fixed regex for GPS
 
 ### Changed
 
 - Better validation for integers
+- Refactor validation in excel for most validators (to include unique & na_ok)